# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Political Website API",
      "description": "政治網站後端 API 文檔",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000/api",
        "description": "開發環境"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
        "Policy": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "政策ID"
            },
            "title": {
              "type": "string",
              "description": "政策標題"
            },
            "description": {
              "type": "string",
              "description": "簡短描述"
            },
            "content": {
              "type": "string",
              "description": "詳細內容"
            },
            "image_url": {
              "type": "string",
              "description": "圖片URL"
            },
            "category_id": {
              "type": "integer",
              "description": "類別ID"
            },
            "objectives": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "objective": {
                    "type": "string",
                    "description": "政策目標"
                  }
                }
              }
            },
            "implementations": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "step": {
                    "type": "string",
                    "description": "實施步驟"
                  },
                  "progress": {
                    "type": "integer",
                    "description": "進度百分比"
                  }
                }
              }
            }
          }
        },
        "Activity": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "活動ID"
            },
            "title": {
              "type": "string",
              "description": "活動標題"
            },
            "description": {
              "type": "string",
              "description": "活動描述"
            },
            "date": {
              "type": "string",
              "format": "date",
              "description": "活動日期"
            },
            "time_start": {
              "type": "string",
              "format": "time",
              "description": "開始時間"
            },
            "time_end": {
              "type": "string",
              "format": "time",
              "description": "結束時間"
            },
            "location": {
              "type": "string",
              "description": "活動地點"
            },
            "category_id": {
              "type": "integer",
              "description": "類別ID"
            },
            "image_url": {
              "type": "string",
              "description": "圖片URL"
            },
            "external_link": {
              "type": "string",
              "description": "外部鏈接"
            },
            "agenda": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "time_slot": {
                    "type": "string",
                    "description": "時間段"
                  },
                  "description": {
                    "type": "string",
                    "description": "議程描述"
                  }
                }
              }
            }
          }
        },
...
      }
    },
    "paths": {
...
      "/policies": {
        "get": {
          "summary": "獲取所有政策",
          "tags": ["Policies"],
          "responses": {
            "200": {
              "description": "成功獲取政策列表",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Policy"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "創建新政策",
          "tags": ["Policies"],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "category_id": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string",
                      "format": "binary"
                    },
                    "objectives": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "objective": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "政策創建成功"
            },
            "400": {
              "description": "請求數據驗證失敗"
            },
            "401": {
              "description": "未授權"
            }
          }
        }
      },
...
      "/activities": {
        "get": {
          "summary": "獲取所有活動",
          "tags": ["Activities"],
          "responses": {
            "200": {
              "description": "成功獲取活動列表",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Activity"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "創建新活動",
          "tags": ["Activities"],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "活動創建成功"
            }
          }
        }
      },
...
    }
  }